+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/time/TimeTaskViewModel.kt
@@ -20,18 +20,10 @@ import com.google.android.ground.ui.datacollection.tasks.AbstractTaskViewModel
 import java.text.DateFormat
 import java.util.Date
 import javax.inject.Inject
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 
 class TimeTaskViewModel @Inject constructor() : AbstractTaskViewModel() {
 
-  private var _timeText: MutableStateFlow<String> = MutableStateFlow("")
-  val timeText = _timeText.asStateFlow()
-
   fun updateResponse(dateFormat: DateFormat?, date: Date) {
-    dateFormat?.let {
-      _timeText.value = it.format(date)
-      setValue(fromMillis(date.time))
-    }
+    dateFormat?.let { setValue(fromMillis(date.time)) }
   }
 }
